[MTS][123233931] Fix SVG bug
https://www.pivotaltracker.com/story/show/123233931

In the MS Edge browser, the event tracker was picking up an object that
wasn't a normal HTML element. It was an SVGElementInstance object. It
did not have a getAttribute() method or an attributes property.

While the addition of a check for the existence of a getAttribute()
method on the element in getAtts() was helpful, the event tracker still
would not have worked with MS Edge, since getTags() used
element.attributes, and passed this value to Object.keys(). In the case
of an SVGElementInstance object, this would result in Object.keys()
evaluating undefined, and would raise an error.

To circumvent this, we use the object's correspondingElement property
higher up in the call stack as the node. For an SVGElementInstance, this
will be an HTML element that will work with the rest of the event
tracker code. For a normal HTML element, this will result in undefined,
and we end up using the original element as our node.